remove unused, modularise
re-run `pip install -e .` after pull
