+++ pyproject.toml
@@ -20,5 +20,5 @@ dependencies = [
 dynamic = ["version"]
 
 [project.scripts]
-warnet = "warnet.connect_nodes:main"
+warnet = "warnet.main:main"
 warnet-cli = "warnet.warnet_cli:main"
